Product controller functions done.

+++ backend/src/controllers/product.controller.js
@@ -16,7 +16,13 @@ productCtrl.listProductById = async (req, res) => {
   try {
     const { id } = req.params;
     const product = await productModel.findById({ _id: id });
-  } catch (error) {}
+    if(!product){
+      return generalMessage(res, 404, "", false, "Product not found.");
+    }
+    generalMessage(res, 200, product, true, "Data found.");
+  } catch (error) {
+    generalMessage(res, 404, "", false, error.message);
+  }
 };
 
 productCtrl.addProduct = async (req, res) => {
@@ -31,7 +37,7 @@ productCtrl.addProduct = async (req, res) => {
     newProduct.setImgUrl(filename);
     await newProduct.save();
     console.log(newProduct);
-    generalMessage(res, 201, newProduct, true, `Product ${name} added`);
+    generalMessage(res, 201, newProduct, true, `Product "${name}" added`);
   } catch (error) {
     generalMessage(res, 400, "", false, error.message);
   }
@@ -47,19 +53,39 @@ productCtrl.updateProduct = async (req, res) => {
     const name = req.body.name || product.name;
     const description = req.body.description || product.description;
     const price = req.body.price || product.price;
+    if (req.file) {
+      if (product.nameImg) {
+        deleteImg(product.nameImg);
+      }
+      const { filename } = req.file;
+      product.setImgUrl(filename);
+      await product.save();
+      console.log(product.img);
+    }
     const nameImg = product.nameImg;
     const img = product.img;
-    if (req.file) {
-      deleteImg(product.nameImg);
-      const filename = req.file;
-      product.setImgUrl(filename);
-    }
     const updatedProduct = { name, description, price, nameImg, img };
     await product.updateOne(updatedProduct);
-    generalMessage(res, 201, updatedProduct, true, `Product ${name} updated`);
+    generalMessage(res, 201, updatedProduct, true, `Product "${name}" updated`);
   } catch (error) {
-    generalMessage(res, 404, "", false, error.message);
+    generalMessage(res, 500, "", false, error.message);
   }
 };
 
+productCtrl.deleteProduct = async (req, res) => {
+  try {
+    const { id } = req.params;
+    const product = await productModel.findById({ _id: id });
+    if (!product) {
+      return generalMessage(res, 404, "", false, "Product not found.");
+    }
+    if (product.nameImg) {
+      deleteImg(product.nameImg);
+    }
+    await product.deleteOne({ _id: id });
+    generalMessage(res, 202, "", false, `Product "${product.name}" deleted.`);
+  } catch (error) {
+    generalMessage(res, 500, "", true, error.message);
+  }
+};
 module.exports = productCtrl;
